@@ -1,14 +1,15 @@
 {
   "name": "signalk-server",
-  "version": "2.2.0",
+  "version": "2.13.0-beta.2",
   "description": "An implementation of a [Signal K](http://signalk.org) server for boats.",
   "main": "index.js",
   "scripts": {
-    "build": "tsc --build",
-    "build:all": "npm run build:workspaces && npm run build && npm run build-declaration",
+    "build": "tsc --build && npm run build:docs",
+    "build:all": "npm run build:workspaces && npm run build && npm run build-declaration && npm run build:docs",
+    "build:docs": "USER_INFO=$(id -u):$(id -g) docker run --rm -v ${PWD}/docs:/book peaceiris/mdbook:v0.4.40 build",
     "build:workspaces": "npm run build --workspaces --if-present",
     "build-declaration": "tsc --declaration",
-    "watch": "npm run build -- -w",
+    "watch": "tsc --build -w",
     "prettier": "prettier -w src/",
     "lint": "eslint -c .eslintrc.js --ext .ts --ext .js --fix src/",
     "format": "npm run prettier && npm run lint",
@@ -71,13 +72,13 @@
   ],
   "dependencies": {
     "@signalk/course-provider": "^1.0.0",
-    "@signalk/n2k-signalk": "^2.0.0",
+    "@signalk/n2k-signalk": "^3.0.0",
     "@signalk/nmea0183-signalk": "^3.0.0",
-    "@signalk/resources-provider": "^1.0.0",
-    "@signalk/server-admin-ui": "2.1.x",
-    "@signalk/server-api": "2.1.x",
+    "@signalk/resources-provider": "^1.3.0",
+    "@signalk/server-admin-ui": "2.12.x",
+    "@signalk/server-api": "2.6.x",
     "@signalk/signalk-schema": "^1.7.1",
-    "@signalk/streams": "^3.2.0",
+    "@signalk/streams": "^4.4.0",
     "api-schema-builder": "^2.0.11",
     "baconjs": "^1.0.1",
     "bcryptjs": "^2.4.3",
@@ -92,7 +93,6 @@
     "cookie-parser": "^1.4.3",
     "cors": "^2.5.2",
     "debug": "^4.3.3",
-    "devcert": "^1.2.2",
     "dnssd2": "1.0.0",
     "errorhandler": "^1.3.0",
     "express": "^4.10.4",
@@ -116,6 +116,7 @@
     "ncp": "^2.0.0",
     "node-fetch": "^2.6.0",
     "primus": "^7.0.0",
+    "selfsigned": "^2.4.1",
     "semver": "^7.5.4",
     "split": "^1.0.0",
     "stat-mode": "^1.0.0",
@@ -126,12 +127,12 @@
   },
   "optionalDependencies": {
     "@signalk/freeboard-sk": "^2.0.0-beta.3",
+    "@mxtommy/kip": "^2.9.1",
     "@signalk/instrumentpanel": "0.x",
     "@signalk/set-system-time": "^1.2.0",
     "@signalk/signalk-to-nmea0183": "^1.0.0",
     "@signalk/vesselpositions": "^1.0.0",
     "@signalk/udp-nmea-plugin": "^2.0.0",
-    "@signalk/zones": "^1.0.0",
     "signalk-to-nmea2000": "^2.16.0",
     "signalk-n2kais-to-nmea0183": "^1.3.1",
     "mdns": "^2.5.1",
@@ -148,15 +149,13 @@
     "@types/mocha": "^10.0.1",
     "@types/ncp": "^2.0.5",
     "@types/node-fetch": "^2.5.3",
-    "@types/rmfr": "^2.0.1",
     "@types/semver": "^7.1.0",
     "@types/serialport": "^8.0.2",
     "@types/split": "^1.0.0",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/unzipper": "^0.10.5",
     "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
+    "@typescript-eslint/eslint-plugin": "^8.16.0",
     "chai": "^4.0.0",
     "chai-json-equal": "0.0.1",
     "chai-things": "^0.2.0",
@@ -168,11 +167,10 @@
     "lint-staged": "^10.0.3",
     "mocha": "^10.2.0",
     "prettier": "^2.7.1",
-    "rimraf": "^3.0.2",
-    "rmfr": "^2.0.0",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.1",
-    "typedoc": "^0.23.23",
-    "typescript": "^4.8.3"
+    "typedoc": "^0.26.11",
+    "typescript": "5.6.x"
   },
   "funding": "https://opencollective.com/signalk"
 }
