@@ -23,9 +23,10 @@ const {
   updateRequest,
   queryRequest
 } = require('../requestResponse')
-const { putPath } = require('../put')
+const { putPath, deletePath } = require('../put')
 import { createDebug } from '../debug'
 import { JsonWebTokenError, TokenExpiredError } from 'jsonwebtoken'
+import { startEvents, startServerEvents } from '../events'
 const debug = createDebug('signalk-server:interfaces:ws')
 const debugConnection = createDebug('signalk-server:interfaces:ws:connections')
 const Primus = require('primus')
@@ -234,6 +235,10 @@ module.exports = function (app) {
                 processPutRequest(spark, msg)
               }
 
+              if (msg.delete) {
+                processDeleteRequest(spark, msg)
+              }
+
               if (msg.requestId && msg.query) {
                 processReuestQuery(spark, msg)
               }
@@ -330,6 +335,28 @@ module.exports = function (app) {
       msg.context,
       msg.put.path,
       msg.put,
+      spark.request,
+      msg.requestId,
+      (reply) => {
+        debug('sending put update %j', reply)
+        spark.write(reply)
+      }
+    ).catch((err) => {
+      console.error(err)
+      spark.write({
+        requestId: msg.requestId,
+        state: 'COMPLETED',
+        statusCode: 502,
+        message: err.message
+      })
+    })
+  }
+
+  function processDeleteRequest(spark, msg) {
+    deletePath(
+      app,
+      msg.context,
+      msg.delete.path,
       spark.request,
       msg.requestId,
       (reply) => {
@@ -504,18 +531,42 @@ function processUpdates(app, pathSources, spark, msg) {
   })
 }
 
+/*
+  Keep a list of shared context-path strings that is shared across ws connections.
+  This way the string values are shared across ws connections and not recreated
+  for each context-path-ws combination. This reduces memory consumption for
+  multiple ws clients.
+  Nevertheless we need to purge this data eventually, otherwise the strings for 
+  AIS targets will stay forever, so implement a simple total purge. This may cause
+  some thrashing, but is better than not sharing the values.
+*/
+let canonical_meta_contextpath_values = {}
+const getContextPathMetaKey = (context, path) => {
+  const contextPaths =
+    canonical_meta_contextpath_values[context] ||
+    (canonical_meta_contextpath_values[context] = {})
+  const result =
+    contextPaths[path] || (contextPaths[path] = `${context}.${path}`)
+  return result
+}
+setInterval(() => {
+  canonical_meta_contextpath_values = {}
+}, 30 * 60 * 1000)
+
 function handleValuesMeta(kp) {
-  if (kp.path && !this.spark.sentMetaData[this.context + '.' + kp.path]) {
+  const fullContextPathKey = getContextPathMetaKey(this.context, kp.path)
+  if (kp.path && !this.spark.sentMetaData[fullContextPathKey]) {
     const split = kp.path.split('.')
     for (let i = split.length; i > 1; i--) {
       const path = split.slice(0, i).join('.')
-      if (this.spark.sentMetaData[this.context + '.' + path]) {
+      const partialContextPathKey = getContextPathMetaKey(this.context, path)
+      if (this.spark.sentMetaData[partialContextPathKey]) {
         //stop backing up the path with first prefix that has already been handled
         break
       } else {
         //always set to true, even if there is no meta for the path
-        this.spark.sentMetaData[this.context + '.' + path] = true
-        let meta = getMetadata(this.context + '.' + path)
+        this.spark.sentMetaData[partialContextPathKey] = true
+        let meta = getMetadata(partialContextPathKey)
         if (meta) {
           this.spark.write({
             context: this.context,
@@ -678,7 +729,20 @@ function handleRealtimeConnection(app, spark, onChange) {
 
   if (spark.query.serverevents === 'all') {
     spark.hasServerEvents = true
-    startServerEvents(app, spark)
+    startServerEvents(
+      app,
+      spark,
+      wrapWithverifyWS(app.securityStrategy, spark, spark.write.bind(spark))
+    )
+  }
+
+  if (spark.query.events) {
+    startEvents(
+      app,
+      spark,
+      wrapWithverifyWS(app.securityStrategy, spark, spark.write.bind(spark)),
+      spark.query.events
+    )
   }
 }
 
@@ -696,49 +760,6 @@ function sendLatestDeltas(app, deltaCache, selfContext, spark) {
       sendMetaData(app, spark, delta)
       spark.write(delta)
     })
-}
-
-function startServerEvents(app, spark) {
-  const onServerEvent = wrapWithverifyWS(
-    app.securityStrategy,
-    spark,
-    spark.write.bind(spark)
-  )
-  app.on('serverevent', onServerEvent)
-  spark.onDisconnects.push(() => {
-    app.removeListener('serverevent', onServerEvent)
-  })
-  try {
-    spark.write({
-      type: 'VESSEL_INFO',
-      data: {
-        name: app.config.vesselName,
-        mmsi: app.config.vesselMMSI,
-        uuid: app.config.vesselUUID
-      }
-    })
-  } catch (e) {
-    if (e.code !== 'ENOENT') {
-      console.error(e)
-    }
-  }
-  Object.keys(app.lastServerEvents).forEach((propName) => {
-    spark.write(app.lastServerEvents[propName])
-  })
-  spark.write({
-    type: 'DEBUG_SETTINGS',
-    data: app.logging.getDebugSettings()
-  })
-  if (app.securityStrategy.canAuthorizeWS()) {
-    spark.write({
-      type: 'RECEIVE_LOGIN_STATUS',
-      data: app.securityStrategy.getLoginStatus(spark.request)
-    })
-  }
-  spark.write({
-    type: 'SOURCEPRIORITIES',
-    data: app.config.settings.sourcePriorities || {}
-  })
 }
 
 function startServerLog(app, spark) {
@@ -765,6 +786,10 @@ function startServerLog(app, spark) {
 function getAssertBufferSize(config) {
   const MAXSENDBUFFERSIZE =
     process.env.MAXSENDBUFFERSIZE || config.maxSendBufferSize || 4 * 512 * 1024
+  const MAXSENDBUFFERCHECKTIME =
+    process.env.MAXSENDBUFFERCHECKTIME ||
+    config.maxSendBufferCheckTime ||
+    30 * 1000
   debug(`MAXSENDBUFFERSIZE:${MAXSENDBUFFERSIZE}`)
 
   if (MAXSENDBUFFERSIZE === 0) {
@@ -774,10 +799,23 @@ function getAssertBufferSize(config) {
   return (spark) => {
     debug(spark.id + ' ' + spark.request.socket.bufferSize)
     if (spark.request.socket.bufferSize > MAXSENDBUFFERSIZE) {
-      spark.end({
-        errorMessage: 'Server outgoing buffer overflow, terminating connection'
-      })
-      console.error('Send buffer overflow, terminating connection ' + spark.id)
+      if (!spark.bufferSizeExceeded) {
+        console.warn(
+          `${spark.id} outgoing buffer > max:${spark.request.socket.bufferSize}`
+        )
+        spark.bufferSizeExceeded = Date.now()
+      }
+      if (Date.now() - spark.bufferSizeExceeded > MAXSENDBUFFERCHECKTIME) {
+        spark.end({
+          errorMessage:
+            'Server outgoing buffer overflow, terminating connection'
+        })
+        console.error(
+          'Send buffer overflow, terminating connection ' + spark.id
+        )
+      }
+    } else {
+      spark.bufferSizeExceeded = undefined
     }
   }
 }
